Create PaymentProcessor module.
