+++ lib/js-Round1.js
@@ -8,41 +8,7 @@
 
 'use strict';
 
-var BraintreePayment = require('./BraintreePayment.js');
-var PaypalPayment = require('./PaypalPayment.js');
-
-
-var paymentProcessor = {
-  // TODO: Enable 'create' from paypal api
-  processPaymentRequest: function(create_payment_json, reply) {
-    // TODO: handle malformed data
-    // TODO: this logic is confusing and error prone, cleanup
-    var currency = create_payment_json.transactions[0].amount.currency;
-    var payment = new PaypalPayment();
-    if(create_payment_json.payer.funding_instruments[0].credit_card.type === "amex" ) {
-      if( currency === "USD" ) {
-        payment.createPaypalPayment(create_payment_json, reply);
-      }
-      else {
-        process.nextTick( function() {
-          var error = new Error("American Express must be used with US Dollars only.");
-          reply(error,"American Express must be used with US Dollars only.");
-        });
-      }
-    }
-    else {
-      if( currency === "USD" || currency === "EUR" || currency === "AUD" ) {
-        payment.createPaypalPayment(create_payment_json, reply);
-      }
-      else {
-        var braintreePay = new BraintreePayment();
-        braintreePay.createBraintreePayment(create_payment_json, reply);
-      }
-    }
-  }
-};
-
-exports.paymentProcessor = paymentProcessor;
+var PaymentProcessor = require('./PaymentProcessor.js');
 
 var paymentServer = {
   create_listener: function(basePath) {
@@ -74,11 +40,12 @@ var paymentServer = {
       }
     });
 
+    var processor = new PaymentProcessor();
     server.route({
       // TODO: remove GET from list
       method: ['GET','PUT', 'POST'],
       path: '/ccpay',
-      handler: paymentProcessor.processPaymentRequest
+      handler: processor.processPaymentRequest
     });
 
     return server;
